student/requirements.txt
Django==2.1
django-tastypie==0.14.2
psycopg2
